requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication and OAuth
django-allauth==0.61.1
google-auth==2.28.0

# Database and ORM
psycopg2-binary==2.9.9
django-environ==0.21.0

# Task Queue
celery==5.3.4
redis==5.0.1

# API Documentation
drf-spectacular==0.27.0

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-django==0.42b0
opentelemetry-instrumentation-celery==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-otlp==0.42b0
opentelemetry-exporter-otlp-proto-http==0.42b0
python-json-logger==2.0.7

# Web Server
gunicorn==21.2.0
whitenoise==6.6.0

# Financial Calculations
numpy==1.26.3
pandas==2.1.4
scipy==1.12.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.13.0
pytest-asyncio==0.23.2
pytest-playwright==0.4.3
factory-boy==3.3.0

# UI Testing
playwright==1.41.2

# Code Quality
flake8==7.0.0
black==24.1.1
isort==5.13.2

# Serialization and Validation
pydantic==2.5.3

# Timezone and Date Handling
pytz==2024.1

# PDF Generation (Phase 2)
reportlab==4.0.7

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Environment
python-decouple==3.8
